+++ backend/moa/src/main/java/com/ssafy/moa/api/repository/BuddyRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BuddyRepository extends JpaRepository<Buddy, Integer> {
+public interface BuddyRepository extends JpaRepository<Buddy, Long> {
 
 }
