+++ backend/moa/src/main/java/com/ssafy/moa/api/repository/MemberRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface MemberRepository extends JpaRepository<Member, Integer> {
-    Optional<Member> findByMemberId(int memberId);
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    Optional<Member> findByMemberId(long memberId);
 }
